+++ src/controller/java/src/matter/controller/CompletionListenerAdapter.kt
@@ -29,17 +29,17 @@ class CompletionListenerAdapter(val listener: MatterController.CompletionListene
 
   override fun onStatusUpdate(status: Int) = listener.onStatusUpdate(status)
 
-  override fun onPairingComplete(errorCode: Int) = listener.onPairingComplete(errorCode)
+  override fun onPairingComplete(errorCode: Long) = listener.onPairingComplete(errorCode.toUInt())
 
-  override fun onPairingDeleted(errorCode: Int) = listener.onPairingDeleted(errorCode)
+  override fun onPairingDeleted(errorCode: Long) = listener.onPairingDeleted(errorCode.toUInt())
 
   override fun onNotifyChipConnectionClosed() = listener.onNotifyChipConnectionClosed()
 
-  override fun onCommissioningComplete(nodeId: Long, errorCode: Int) =
-    listener.onCommissioningComplete(nodeId, errorCode)
+  override fun onCommissioningComplete(nodeId: Long, errorCode: Long) =
+    listener.onCommissioningComplete(nodeId, errorCode.toUInt())
 
-  override fun onCommissioningStatusUpdate(nodeId: Long, stage: String?, errorCode: Int) =
-    listener.onCommissioningStatusUpdate(nodeId, stage, errorCode)
+  override fun onCommissioningStatusUpdate(nodeId: Long, stage: String?, errorCode: Long) =
+    listener.onCommissioningStatusUpdate(nodeId, stage, errorCode.toUInt())
 
   override fun onReadCommissioningInfo(
     vendorId: Int,
@@ -52,10 +52,10 @@ class CompletionListenerAdapter(val listener: MatterController.CompletionListene
 
   override fun onICDRegistrationInfoRequired() = listener.onICDRegistrationInfoRequired()
 
-  override fun onICDRegistrationComplete(errorCode: Int, icdDeviceInfo: ICDDeviceInfo) =
+  override fun onICDRegistrationComplete(errorCode: Long, icdDeviceInfo: ICDDeviceInfo) =
     listener.onICDRegistrationComplete(errorCode, icdDeviceInfo)
 
-  override fun onError(error: Throwable) = listener.onError(error)
+  override fun onError(error: Throwable) = listener.onError(error.toUInt())
 
   override fun onCloseBleComplete() {
     logger.log(Level.INFO, "Not implemented, override the abstract function.")
